refactor: change back to camelCase files and fix linting errors
