+++ package.json
@@ -7,6 +7,7 @@
     "typings": "typings",
     "release": "rimraf dist && tsc && npm t && cp package.json README.md dist/src && ts-node build/links.ts --src ./dist/src",
     "test": "rimraf dist && tsc && mocha dist/test/**",
+    "lint": "tslint -c tslint.json src/**/*.ts test/**/*.ts",
     "test:watch": "npm run test -- -w",
     "test:watch:debug": "npm run test -- -w --debug-brk=5905",
     "tsc": "tsc",
