+++ src/directiveClassSuffixRule.ts
@@ -1,20 +1,19 @@
 import * as Lint from 'tslint/lib/lint';
 import * as ts from 'typescript';
 import {sprintf} from 'sprintf-js';
-import SyntaxKind = require('./util/syntax-kind');
-import {Ng2Walker} from './angular/ng2-walker';
+import {Ng2Walker} from './angular/ng2Walker';
 
 export class Rule extends Lint.Rules.AbstractRule {
+  static FAILURE:string = 'The name of the class %s should end with the suffix Directive ($$02-03$$)';
+
+  static validate(className:string):boolean {
+    return /.*Directive/.test(className);
+  }
+
   public apply(sourceFile:ts.SourceFile):Lint.RuleFailure[] {
     return this.applyWithWalker(
       new ClassMetadataWalker(sourceFile,
         this.getOptions()));
-  }
-
-  static FAILURE:string = "The name of the class %s should end with the suffix Directive ($$02-03$$)";
-
-  static validate(className:string):boolean {
-    return /.*Directive/.test(className);
   }
 }
