+++ src/noAttributeParameterDecoratorRule.ts
@@ -1,30 +1,29 @@
 import * as Lint from 'tslint/lib/lint';
 import * as ts from 'typescript';
 import {sprintf} from 'sprintf-js';
-import SyntaxKind = require('./util/syntax-kind');
+import SyntaxKind = require('./util/syntaxKind');
 
 export class Rule extends Lint.Rules.AbstractRule {
-  public apply(sourceFile:ts.SourceFile):Lint.RuleFailure[] {
-    return this.applyWithWalker(
-      new ConstructorMetadataWalker(sourceFile,
-        this.getOptions()));
-  }
-
   static  FAILURE_STRING:string = 'In the constructor of class "%s",' +
     ' the parameter "%s" uses the @Attribute decorator, ' +
     'which is considered as a bad practice. Please,' +
     ' consider construction of type "@Input() %s: string"';
 
+  public apply(sourceFile:ts.SourceFile):Lint.RuleFailure[] {
+    return this.applyWithWalker(
+      new ConstructorMetadataWalker(sourceFile,
+        this.getOptions()));
+  }
 }
 
 export class ConstructorMetadataWalker extends Lint.RuleWalker {
   visitConstructorDeclaration(node:ts.ConstructorDeclaration) {
     let syntaxKind = SyntaxKind.current();
-    let parentName:string="";
+    let parentName: string = '';
     let parent = (<any>node.parent);
-    if(parent.kind===syntaxKind.ClassExpression){
+    if (parent.kind === syntaxKind.ClassExpression) {
       parentName= parent.parent.name.text;
-    }else if(parent.kind = syntaxKind.ClassDeclaration){
+    } else if (parent.kind = syntaxKind.ClassDeclaration) {
       parentName= parent.name.text;
     }
     (node.parameters || []).forEach(this.validateParameter.bind(this, parentName));
