+++ src/noForwardRefRule.ts
@@ -1,18 +1,18 @@
 import * as Lint from 'tslint/lib/lint';
 import * as ts from 'typescript';
 import {sprintf} from 'sprintf-js';
-import SyntaxKind = require('./util/syntax-kind');
+import SyntaxKind = require('./util/syntaxKind');
 
 export class Rule extends Lint.Rules.AbstractRule {
+  static FAILURE_IN_CLASS: string = 'Avoid using forwardRef in class "%s"';
+
+  static FAILURE_IN_VARIABLE: string = 'Avoid using forwardRef in variable "%s"';
+
   public apply(sourceFile:ts.SourceFile):Lint.RuleFailure[] {
     return this.applyWithWalker(
       new ExpressionCallMetadataWalker(sourceFile,
         this.getOptions()));
   }
-
-  static FAILURE_IN_CLASS: string = 'Avoid using forwardRef in class "%s"';
-
-  static FAILURE_IN_VARIABLE: string = 'Avoid using forwardRef in variable "%s"';
 }
 
 export class ExpressionCallMetadataWalker extends Lint.RuleWalker {
@@ -28,10 +28,10 @@ export class ExpressionCallMetadataWalker extends Lint.RuleWalker {
         currentNode = currentNode.parent;
       }
       let failureConfig:string[] =[];
-      if(currentNode.kind===SyntaxKind.current().VariableStatement){
-        failureConfig=[Rule.FAILURE_IN_VARIABLE,currentNode.declarationList.declarations[0].name.text];
-      }else{
-        failureConfig=[Rule.FAILURE_IN_CLASS,currentNode.name.text];
+      if (currentNode.kind===SyntaxKind.current().VariableStatement) {
+        failureConfig = [Rule.FAILURE_IN_VARIABLE,currentNode.declarationList.declarations[0].name.text];
+      } else {
+        failureConfig = [Rule.FAILURE_IN_CLASS,currentNode.name.text];
       }
       this.addFailure(
         this.createFailure(
