+++ src/util/selectorValidator.ts
