+++ src/util/utils.ts
@@ -20,4 +20,3 @@ export const stringDistance = (s: string, t: string, ls: number = s.length, lt: 
   }
   return memo[lt];
 };
-
